PycharmProjects/Autotests/shop-tests/pyproject.toml
[tool.poetry]
name = "befree-tests"
version = "0.1.0"
description = ""
authors = ["potegovaav <potegovaav@melonfashion.com>"]
readme = "README.md"
packages = [{ include = "befree_tests" }]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
pytest = "7.2.0"
requests = "2.31.0"
voluptuous = "^0.13.1"
pytest-voluptuous = "^1.2.0"
selene = "2.0.0rc9"
allure-pytest = "^2.13.2"
python-dotenv = ">=0.21.0,<0.22.0"
pytest-check = "^2.1.5"
mimesis = "^10.1.0"
curlify = "^2.2.1"
xmltodict = "^0.13.0"
pytest-soft-assertions = "^0.1.2"
pandas = "^2.2.3"
psycopg2-binary = "^2.9.10"
charset-normalizer	= "^3.4.2"
numpy =	"^2.2.6"
ruff = "^0.11.11"
pytest-playwright = "^0.5.2"
playwright = "^1.51.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = [
    "--alluredir=allure-results",
    "--allure-no-capture",
    "--headed"
]
